requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1
boto3==1.36.2
openpyxl==3.1.5
plotly==6.0.0
